pyproject.toml
[tool.poetry]
name = "mppw"
version = "0.1.0"
description = ""
authors = ["ASCC <ascc@maine.edu>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.68.2"
motor = "^2.5.1"
argh = "^0.26.2"
uvicorn = {extras = ["standard"], version = "^0.15.0"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.5"
alembic = "^1.7.4"
sqlmodel = "^0.0.4"
psycopg2 = "^2.9.1"
furl = "^2.1.3"
toml = "^0.10.2"
docker-compose = "^1.29.2"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"

[tool.poetry.scripts]
docker-build = "scripts.docker_build:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
